requirements.txt
googlesearch-python>=1.2
pandas>=2.0
openpyxl>=3.1